% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfTools.r
\name{cdsfy}
\alias{cdsfy}
\title{gets features that appear in the mRNA between the start and stop codon (inclusive) in transcription order}
\usage{
cdsfy(gtf)
}
\description{
gets features that appear in the mRNA between the start and stop codon (inclusive) in transcription order
}
