% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfTools.r
\name{collapse2tx}
\alias{collapse2tx}
\title{collapse all exons and introns of transcripts to 1 annotation}
\usage{
collapse2tx(gtf, trcol)
}
\description{
collapse all exons and introns of transcripts to 1 annotation
}
