% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfTools.r
\name{pvarAnnotate}
\alias{pvarAnnotate}
\title{given a gtf with the neccessary columns, it gets the variant classification of every possible mutation in the gtf's transcripts}
\usage{
pvarAnnotate(gtf, .genome)
}
\description{
given a gtf with the neccessary columns, it gets the variant classification of every possible mutation in the gtf's transcripts
}
