% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfTools.r
\name{pvarCDSannotate}
\alias{pvarCDSannotate}
\title{assumes all transcripts have been checked to be divisible by 3!}
\usage{
pvarCDSannotate(cdsgtf, genome)
}
\description{
assumes all transcripts have been checked to be divisible by 3!
}
