% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfTools.r
\name{rnaify}
\alias{rnaify}
\title{gets features that get transcribed to RNA in transcription order}
\usage{
rnaify(gtf, mrnaOnly = FALSE)
}
\description{
gets features that get transcribed to RNA in transcription order
}
